Use a replacement token for the per-version documentation root
If a documentation path contains $(DOC_BASE_FOR_VERSION) we replace
this with `/documentation/$(version)`

Why

Previously, if the path did not start with "http://" or "https://" or
"/", we inferred that it is relative to the per-version documentation
dir. A path of `html/123` for version 0.13.0 would be expanded to:
documentation/0.13.0/html/123.

Using a replacement names this concept and trims down the templating
code to not need to check the path prefixes.

Signed-off-by: Robert Young <robeyoun@redhat.com>