Enable github forks to override jekyll config
The github build action will dump a `JEKYLL_CONFIG_OVERRIDES` action
variable into a file `_config-overrides.yml` which will override
whatever you want in `_config.yml`.

Why:

It's convenient to be able to deploy the website on a forked repository.
Currently this doesn't work smoothly because by default with actions
because it builds with kroxylicious.io as the url. The mechanism that
configures the baseurl is also not wired up correctly.

To override the url we need to add in an override configuration, so I
figure we should use the same mechanism to override baseurl. The fork
can set a variable containing both these values.

Signed-off-by: Robert Young <robeyoun@redhat.com>