Add Development Mode warning.
Adds a transparent warning to the top right of the page when in
development environment. This is the default jekyll environment so in
the action I'm setting it to production for the real deal, though
nothing acts on the "production" value currently.

Swaps the favicon for a grayscale version to distinguish it.

We are considering deploying a development verstion of the website
continuously to GitHub pages somewhere, so we need a way to disambiguate
from the main page in case someone stumbles on it. It will also make it
immediately clear when developing locally that we are looking at a dev
deployment.

Signed-off-by: Robert Young <robeyoun@redhat.com>