Use root relative redirect paths and make absolute in layout
Why:
I want to re-use the redirect layout from some HTML frontmatter, and
you cannot use liquid templating nicely in the Jekyll frontmatter to
create an absolute url like "target: {{"/path/to/target" | absolute_url}}".
I want to use root relative paths so that it works when there is a base
url.

In future the layout would need adjustment if we wanted it to handle
absolute redirect targets, maybe passing a different property to the
page, or sniffing the first N characters for http:/https:

Signed-off-by: Robert Young <robertyoungnz@gmail.com>