Add /get-started which redirects to the latest quickstart
Why: I want a permanent and snappy URL that redirects to the latest
proxy quick start so that I can embed it in videos etc. This should be
long lasting so I've added simple tests (invoked at the end of build.sh)
to check that the page, target page and link are what we expect.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>